requirements.txt
pandas
reportlab
python-dotenv
Pillow
